Move ffdata.h back into src/
