+++ src/finfield.c
@@ -83,7 +83,7 @@
 #include	"hpc/thread.h"		/* threads			   */
 #include	"hpc/tls.h"			/* thread-local storage		   */
 
-#include	"hpc/ffdata.h"		/* pre-computed finite field data  */
+#include	"ffdata.h"		/* pre-computed finite field data  */
 
 
 /****************************************************************************
